Add common api utils
